feat: replacing tinyint because not existing in postgres

+++ db/entity/command.ts
@@ -8,7 +8,7 @@ export class CommandEntity extends BaseEntity {
   @Column('int', { default: 0 })
   counter: number
 
-  @Column('tinyint')
+  @Column('int')
   permissions: number
 
   @Column('varchar')
